UniValue: Add a `static_assert` to the `setInt64` private function
Summary
---

The `UniValue::setInt64` function makes an assumption that it is always
called with either an `int64_t` or `uint64_t` argument. If the caller
violates this assumption, it's possible to get UB.

Since this is a potential hazard for future maintainers, and in the
interests of code quality, it's safer to enforce this assumption at
compile-time via a `static_assert`. This commit adds such a
`static_assert`, enforcing that this function only be called with a
`*int64_t` argument.

This commit is entirely a code quality commit and no behavior is changed
(in fact not even emitted assembly code should change as a result of
this commit).

Test Plan
---

- `ninja all check-all`

+++ src/univalue/include/univalue.h
@@ -744,9 +744,9 @@ private:
     static void stringify(Stream & stream, const UniValue::Array& value, unsigned int prettyIndent, unsigned int indentLevel);
     static void stringify(Stream & stream, const std::string& value, unsigned int prettyIndent, unsigned int indentLevel);
 
-    // Used by the setInt() overloads
-    template<typename Integer>
-    void setInt64(Integer val);
+    // Used internally by the integral operator=() overloads
+    template<typename Int64>
+    void setInt64(Int64 val);
 
 public:
     // Strict type-specific getters, these throw std::runtime_error if the
